Share logic with guard.js
