+++ src/index.js
@@ -1,4 +1,5 @@
-import { guard, coerce } from "./guard.js";
-import { inferEnv, inferType } from "./infer.js";
+import { guard } from "./guard.js";
+import { inferEnv } from "./infer.js";
+import { coerce, inferType, isBoolean, isNumeric } from "./utils.js";
 
-export { guard, coerce, inferEnv, inferType };
+export { guard, inferEnv, coerce, inferType, isBoolean, isNumeric };
